[claude] Audit remediation Tiers 2-3: code fixes, C++20 upgrade, deviation review
Tier 2 code fixes (B1-B6):
- B1: Split multi-expression lines (main.cpp, ws2812_status.c)
- B2: One variable per declaration (main.cpp struct + locals)
- B3: Remove continue where possible (i2c_bus.c), keep 2 justified
- B4: Fixed-width loop counters across 7 files (int → uint8_t/int32_t)
- B5: Pointer style consistency (Type* name) across headers
- B6: Add -Werror globally, -Wpedantic for project sources only

Tier 3 deviation review (C16-C22):
- C16: C++20 upgrade (constexpr debug macros, volatile loop fix)
- C17: stdio.h mitigated via runtime lockout plan + MISRA deviation
- C20: C++20 standard set in CMakeLists.txt
- Items #15,#18,#19,#21,#22 dispositioned (accepted/deferred/pass)

Documentation:
- Code classification framework added to CODING_STANDARDS.md
- Matek M8Q-5883 GPS added to HARDWARE.md with u-blox migration path
- Full remediation tracking in AUDIT_REMEDIATION.md
- Audit dashboard updated (82% → 89% projected)

Build: 103/103 clean, -Werror -Wpedantic enforced, 0 warnings, C++20

+++ src/drivers/icm20948.c
@@ -240,7 +240,7 @@ static bool init_magnetometer(icm20948_t* dev) {
 
     // 5. Verify WHO_AM_I with retry (per SparkFun: up to 10 attempts)
     bool mag_found = false;
-    for (int tries = 0; tries < 10; tries++) {
+    for (uint8_t tries = 0; tries < 10; tries++) {
         uint8_t wia2;
         if (mag_read_reg(dev, AK_WIA2, &wia2) && wia2 == AK09916_WHO_AM_I) {
             mag_found = true;
@@ -315,7 +315,7 @@ bool icm20948_init(icm20948_t* dev, uint8_t addr) {
 
     // Initialize magnetometer
     // Initialize magnetometer with retries (intermittent after reboot)
-    for (int mag_attempt = 0; mag_attempt < 3; mag_attempt++) {
+    for (uint8_t mag_attempt = 0; mag_attempt < 3; mag_attempt++) {
         if (init_magnetometer(dev)) {
             break;
         }
